add lib.py for internal usage
